Replace `half3` with `float3` in shaders

+++ gamedata/shaders/r1/common.h
@@ -12,7 +12,7 @@ uniform float4x4	m_plmap_xform;
 uniform float4 		m_plmap_clamp	[2];	// 0.w = factor
 
 half  	calc_fogging 	(float4 w_pos)	{  return saturate(dot(w_pos,fog_plane)); 	}
-half2 	calc_detail 	(half3 w_pos)	{ 
+half2 	calc_detail 	(float3 w_pos)	{ 
 	float  	dtl	= distance(w_pos,eye_position)*dt_params.w;
 		dtl	= min(dtl*dtl, 1);
 	half  	dt_mul	= 1  - dtl;	// dt*  [1 ..  0 ]
@@ -115,8 +115,8 @@ float3	v_hemi 		(float3 n)		{	return L_hemi_color/* *(.5f + .5f*n.y) */; 		}
 float3	v_hemi_wrap	(float3 n, float w)	{	return L_hemi_color/* *(w + (1-w)*n.y) */; 		}
 float3 	v_sun 		(float3 n)		{	return L_sun_color*max(0,dot(n,-L_sun_dir_w));		}
 float3 	v_sun_wrap	(float3 n, float w)	{	return L_sun_color*(w+(1-w)*dot(n,-L_sun_dir_w));	}
-half3	p_hemi		(float2 tc) 	{
-	//half3	t_lmh 	= tex2D		(s_hemi, tc);
+float3	p_hemi		(float2 tc) 	{
+	//float3	t_lmh 	= tex2D		(s_hemi, tc);
 	//return  dot	(t_lmh,1.h/3.h);
 	float4	t_lmh 	= tex2D		(s_hemi, tc);
 	return  t_lmh.a;
