+++ gamedata/shaders/r1/shared/common.h
@@ -11,19 +11,19 @@ uniform float4		timers;
 uniform float4		fog_plane;
 uniform float4		fog_params;		// x=near*(1/(far-near)), ?,?, w = -1/(far-near)
 uniform float4		fog_color;
-uniform half3		L_sun_color;
-uniform half3		L_sun_dir_w;
-uniform half3		L_sun_dir_e;
+uniform float3		L_sun_color;
+uniform float3		L_sun_dir_w;
+uniform float3		L_sun_dir_e;
 uniform float4		L_hemi_color;
 uniform float4		L_ambient;		// L_ambient.w = skynbox-lerp-factor
 uniform float3 		eye_position;
-uniform half3		eye_direction;
-uniform half3		eye_normal;
+uniform float3		eye_direction;
+uniform float3		eye_normal;
 uniform	float4 		dt_params;
 
-half3 	unpack_normal	(half3 v)	{ return 2*v-1;			}
-half3 	unpack_bx2	(half3 v)	{ return 2*v-1; 		}
-half3 	unpack_bx4	(half3 v)	{ return 4*v-2; 		}
+float3 	unpack_normal	(float3 v)	{ return 2.0f*v-1.0f;			}
+float3 	unpack_bx2	(float3 v)	{ return 2.0f*v-1.0f; 		}
+float3 	unpack_bx4	(float3 v)	{ return 4.0f*v-2.0f; 		}
 
 float2 	unpack_tc_base	(float2 tc, float du, float dv)		{
 		return (tc.xy + float2	(du,dv))*(32.f/32768.f);
