+++ gamedata/shaders/r2/lmodel.h
@@ -5,36 +5,36 @@
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // Lighting formulas			// 
-float4 	plight_infinity		(half m, half3 point, half3 normal, half3 light_direction)       				{
-  half3 N		= normal;							// normal 
-  half3 V 		= -normalize	(point);					// vector2eye
-  half3 L 		= -light_direction;						// vector2light
-  half3 H		= normalize	(L+V);						// half-angle-vector 
-  return tex3D 		(s_material,	half3( dot(L,N), dot(H,N), m ) );		// sample material
+float4 	plight_infinity		(half m, float3 point, float3 normal, float3 light_direction)       				{
+  float3 N		= normal;							// normal 
+  float3 V 		= -normalize	(point);					// vector2eye
+  float3 L 		= -light_direction;						// vector2light
+  float3 H		= normalize	(L+V);						// half-angle-vector 
+  return tex3D 		(s_material,	float3( dot(L,N), dot(H,N), m ) );		// sample material
 }
-float4 	plight_infinity2	(half m, half3 point, half3 normal, half3 light_direction)       				{
-  	half3 N		= normal;							// normal 
-  	half3 V 	= -normalize		(point);		// vector2eye
-  	half3 L 	= -light_direction;					// vector2light
- 	half3 H		= normalize			(L+V);			// half-angle-vector 
-	half3 R     = reflect         	(-V,N);
+float4 	plight_infinity2	(half m, float3 point, float3 normal, float3 light_direction)       				{
+  	float3 N		= normal;							// normal 
+  	float3 V 	= -normalize		(point);		// vector2eye
+  	float3 L 	= -light_direction;					// vector2light
+ 	float3 H		= normalize			(L+V);			// half-angle-vector 
+	float3 R     = reflect         	(-V,N);
 	half 	s	= saturate(dot(L,R));
 			s	= saturate(dot(H,N));
 	half 	f 	= saturate(dot(-V,R));
 			s  *= f;
-	float4	r	= tex3D 			(s_material,	half3( dot(L,N), s, m ) );	// sample material
+	float4	r	= tex3D 			(s_material,	float3( dot(L,N), s, m ) );	// sample material
 			r.w	= pow(saturate(s),4);
   	return	r	;
 }
-float4 	plight_local		(half m, half3 point, half3 normal, half3 light_position, half light_range_rsq, out float rsqr)  {
-  half3 N		= normal;							// normal 
-  half3 L2P 	= point-light_position;                         		// light2point 
-  half3 V 		= -normalize	(point);					// vector2eye
-  half3 L 		= -normalize	((half3)L2P);					// vector2light
-  half3 H		= normalize	(L+V);						// half-angle-vector
+float4 	plight_local		(half m, float3 point, float3 normal, float3 light_position, half light_range_rsq, out float rsqr)  {
+  float3 N		= normal;							// normal 
+  float3 L2P 	= point-light_position;                         		// light2point 
+  float3 V 		= -normalize	(point);					// vector2eye
+  float3 L 		= -normalize	((float3)L2P);					// vector2light
+  float3 H		= normalize	(L+V);						// half-angle-vector
 		rsqr	= dot		(L2P,L2P);					// distance 2 light (squared)
   half  att 	= saturate	(1 - rsqr*light_range_rsq);			// q-linear attenuate
-  float4 light	= tex3D		(s_material, half3( dot(L,N), dot(H,N), m ) ); 	// sample material
+  float4 light	= tex3D		(s_material, float3( dot(L,N), dot(H,N), m ) ); 	// sample material
   return att*light;
 }
