+++ gamedata/shaders/r2/mblur.h
@@ -2,7 +2,7 @@
 #define MBLUR_H
 
 #ifndef 	USE_MBLUR
-half3 	mblur	(float2 UV, half3 pos, half3 c_original)	{ return c_original; }
+float3 	mblur	(float2 UV, float3 pos, float3 c_original)	{ return c_original; }
 #else
 #include "common.h"
 
@@ -13,7 +13,7 @@ uniform half2 	m_blur;		// scale_x / 12, scale_y / 12
 #define MBLUR_SAMPLES 	half(12)
 #define MBLUR_CLAMP	half(0.001)
 
-half3 	mblur	(float2 UV, half3 pos, half3 c_original)	{
+float3 	mblur	(float2 UV, float3 pos, float3 c_original)	{
 	float4 	pos4		= float4	(pos,1.0f);
 
 	float4 	p_current	= mul	(m_current,	pos4);
@@ -23,7 +23,7 @@ half3 	mblur	(float2 UV, half3 pos, half3 c_original)	{
 
 	// For each sample, sum up each sample's color in "Blurred" and then divide
 	// to average the color after all the samples are added.
-	half3 	blurred 	= 	c_original	;
+	float3 	blurred 	= 	c_original	;
         	blurred 	+= 	tex2D(s_image, p_velocity * 1.h  + UV).rgb;
 		blurred		+= 	tex2D(s_image, p_velocity * 2.h  + UV).rgb;
 		blurred		+= 	tex2D(s_image, p_velocity * 3.h  + UV).rgb;
